RAG-AZURE-Flow-Guide/embeddings.py
# embeddings.py
import truststore
truststore.inject_into_ssl()
from openai import OpenAI
from azure_config import OPENAI_API_KEY, EMBEDDING_MODEL

client = OpenAI(api_key=OPENAI_API_KEY)

def get_embedding(text):
    response = client.embeddings.create(
        model=EMBEDDING_MODEL,
        input=text
    )
    return response.data[0].embedding