RAG-AZURE-Flow-Guide/requirements.txt
openai
azure-identity
azure-keyvault-secrets
faiss-cpu
streamlit
numpy
PyPDF2